+++ src/LEGO.AsyncAPI/Models/JsonSchema/FalseApiSchema.cs
@@ -1,0 +1,10 @@
+namespace LEGO.AsyncAPI.Models
+{
+    /// <summary>
+    /// An object representing 'false' for properties of AsyncApiSchema that can be false OR a schema.
+    /// </summary>
+    /// <seealso cref="AsyncApiSchema" />
+    public class FalseApiSchema : AsyncApiSchema
+    {
+    }
+}