+++ test/LEGO.AsyncAPI.Tests/Models/AsyncApiSchema_Should.cs
@@ -1,5 +1,9 @@
 ﻿// Copyright (c) The LEGO Group. All rights reserved.
 
+using System.Linq;
+using LEGO.AsyncAPI.Bindings;
+using LEGO.AsyncAPI.Readers;
+
 namespace LEGO.AsyncAPI.Tests.Models
 {
     using System;
@@ -67,6 +71,9 @@ namespace LEGO.AsyncAPI.Tests.Models
                             MaxLength = 15,
                         },
                     },
+                    AdditionalProperties = new FalseApiSchema(),
+                    Items = new FalseApiSchema(),
+                    AdditionalItems = new FalseApiSchema(),
                 },
                 ["property4"] = new AsyncApiSchema
                 {
@@ -78,7 +85,7 @@ namespace LEGO.AsyncAPI.Tests.Models
                             {
                                 ["property6"] = new AsyncApiSchema
                                 {
-                                    Type = SchemaType.Boolean ,
+                                    Type = SchemaType.Boolean,
                                 },
                             },
                         },
@@ -88,6 +95,60 @@ namespace LEGO.AsyncAPI.Tests.Models
                             MinLength = 2,
                         },
                     },
+                    PatternProperties = new Dictionary<string, AsyncApiSchema>()
+                    {
+                        {
+                            "^S_",
+                            new AsyncApiSchema()
+                            {
+                                Type = SchemaType.String,
+                            }
+                        },
+                        {
+                            "^I_", new AsyncApiSchema()
+                            {
+                                Type = SchemaType.Integer,
+                            }
+                        },
+                    },
+                    PropertyNames = new AsyncApiSchema()
+                    {
+                        Pattern = "^[A-Za-z_][A-Za-z0-9_]*$",
+                    },
+                    AdditionalProperties = new AsyncApiSchema
+                    {
+                        Properties = new Dictionary<string, AsyncApiSchema>
+                        {
+                            ["Property8"] = new AsyncApiSchema
+                            {
+                                Type = SchemaType.String | SchemaType.Null,
+                            },
+                        },
+                    },
+                    Items = new AsyncApiSchema
+                    {
+                        Properties = new Dictionary<string, AsyncApiSchema>
+                        {
+                            ["Property9"] = new AsyncApiSchema
+                            {
+                                Type = SchemaType.String | SchemaType.Null,
+                            },
+                        },
+                    },
+                    AdditionalItems = new AsyncApiSchema
+                    {
+                        Properties = new Dictionary<string, AsyncApiSchema>
+                        {
+                            ["Property10"] = new AsyncApiSchema
+                            {
+                                Type = SchemaType.String | SchemaType.Null,
+                            },
+                        },
+                    },
+                },
+                ["property11"] = new AsyncApiSchema
+                {
+                    Const = new AsyncApiString("aSpecialConstant"),
                 },
             },
             Nullable = true,
@@ -368,6 +429,8 @@ components: { }";
   ""title"": ""title1"",
   ""properties"": {
     ""property1"": {
+      ""items"": false,
+      ""additionalItems"": false,
       ""properties"": {
         ""property2"": {
           ""type"": ""integer""
@@ -376,9 +439,30 @@ components: { }";
           ""type"": ""string"",
           ""maxLength"": 15
         }
-      }
+      },
+      ""additionalProperties"": false
     },
     ""property4"": {
+      ""items"": {
+        ""properties"": {
+          ""Property9"": {
+            ""type"": [
+              ""null"",
+              ""string""
+            ]
+          }
+        }
+      },
+      ""additionalItems"": {
+        ""properties"": {
+          ""Property10"": {
+            ""type"": [
+              ""null"",
+              ""string""
+            ]
+          }
+        }
+      },
       ""properties"": {
         ""property5"": {
           ""properties"": {
@@ -391,7 +475,31 @@ components: { }";
           ""type"": ""string"",
           ""minLength"": 2
         }
+      },
+      ""additionalProperties"": {
+        ""properties"": {
+          ""Property8"": {
+            ""type"": [
+              ""null"",
+              ""string""
+            ]
+          }
+        }
+      },
+      ""patternProperties"": {
+        ""^S_"": {
+          ""type"": ""string""
+        },
+        ""^I_"": {
+          ""type"": ""integer""
+        }
+      },
+      ""propertyNames"": {
+        ""pattern"": ""^[A-Za-z_][A-Za-z0-9_]*$""
       }
+    },
+    ""property11"": {
+      ""const"": ""aSpecialConstant""
     }
   },
   ""nullable"": true,
@@ -407,6 +515,102 @@ components: { }";
             actual = actual.MakeLineBreaksEnvironmentNeutral();
             expected = expected.MakeLineBreaksEnvironmentNeutral();
             actual.Should().Be(expected);
+        }
+
+        [Test]
+        public void Deserialize_WithAdvancedSchema_Works()
+        {
+            // Arrange
+            var json = @"{
+  ""title"": ""title1"",
+  ""properties"": {
+    ""property1"": {
+      ""items"": false,
+      ""additionalItems"": false,
+      ""properties"": {
+        ""property2"": {
+          ""type"": ""integer""
+        },
+        ""property3"": {
+          ""type"": ""string"",
+          ""maxLength"": 15
+        }
+      },
+      ""additionalProperties"": false
+    },
+    ""property4"": {
+      ""items"": {
+        ""properties"": {
+          ""Property9"": {
+            ""type"": [
+              ""null"",
+              ""string""
+            ]
+          }
+        }
+      },
+      ""additionalItems"": {
+        ""properties"": {
+          ""Property10"": {
+            ""type"": [
+              ""null"",
+              ""string""
+            ]
+          }
+        }
+      },
+      ""properties"": {
+        ""property5"": {
+          ""properties"": {
+            ""property6"": {
+              ""type"": ""boolean""
+            }
+          }
+        },
+        ""property7"": {
+          ""type"": ""string"",
+          ""minLength"": 2
+        }
+      },
+      ""additionalProperties"": {
+        ""properties"": {
+          ""Property8"": {
+            ""type"": [
+              ""null"",
+              ""string""
+            ]
+          }
+        }
+      },
+      ""patternProperties"": {
+        ""^S_"": {
+          ""type"": ""string""
+        },
+        ""^I_"": {
+          ""type"": ""integer""
+        }
+      },
+      ""propertyNames"": {
+        ""pattern"": ""^[A-Za-z_][A-Za-z0-9_]*$""
+      }
+    },
+    ""property11"": {
+      ""const"": ""aSpecialConstant""
+    }
+  },
+  ""nullable"": true,
+  ""externalDocs"": {
+    ""url"": ""http://example.com/externalDocs""
+  }
+}";
+            var expected = AdvancedSchemaObject;
+
+            // Act
+            var actual = new AsyncApiStringReader().ReadFragment<AsyncApiSchema>(json, AsyncApiVersion.AsyncApi2_0, out var _diagnostics);
+
+            // Assert
+            actual.Should().BeEquivalentTo(expected);
+            _diagnostics.Errors.Should().BeEmpty();
         }
 
         [Test]
@@ -510,7 +714,7 @@ components: { }";
                     { "testD", new AsyncApiSchema { Reference = new AsyncApiReference { Type = ReferenceType.Schema, Id = "testD" } } },
                 },
             })
-            .WithComponent("testB", new AsyncApiSchema() { Description = "test", Type = SchemaType.Boolean  })
+            .WithComponent("testB", new AsyncApiSchema() { Description = "test", Type = SchemaType.Boolean })
             .Build();
 
             var outputString = new StringWriter(CultureInfo.InvariantCulture);